+++ .github/workflows/run-integration-tests.yaml
@@ -15,9 +15,9 @@ jobs:
 
       - name: Build Docker image 🐳
         run: |
-          docker build -t virus-clade-utils .
+          docker build -t cladetime .
 
       - name: run integration tests 🧪
         run: |
-          docker run --entrypoint /bin/bash virus-clade-utils \
+          docker run --entrypoint /bin/bash cladetime \
           -c "pytest -k integration"
