+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 exclude: >
     (?x)(
-        ^src/virus_clade_utils/bin/|
-        ^src/virus_clade_utils/data/
+        ^src/cladetime/bin/|
+        ^src/cladetime/data/
     )
 repos:
 - repo: https://github.com/astral-sh/ruff-pre-commit
