+++ requirements/requirements-dev.txt
@@ -1,13 +1,13 @@
 # This file was autogenerated by uv via the following command:
 #    uv pip compile pyproject.toml --extra dev -o requirements/requirements-dev.txt
 awscli==1.32.116
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 boto3==1.34.116
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   moto
 boto3-stubs==1.35.28
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 botocore==1.34.116
     # via
     #   awscli
@@ -24,20 +24,20 @@ charset-normalizer==3.3.2
     # via requests
 click==8.1.7
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   rich-click
 cloudpathlib==0.18.1
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 colorama==0.4.6
     # via awscli
 coverage==7.5.3
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 cryptography==43.0.1
     # via moto
 docutils==0.16
     # via awscli
 freezegun==1.5.1
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 idna==3.7
     # via requests
 iniconfig==2.0.0
@@ -59,9 +59,9 @@ markupsafe==2.1.5
 mdurl==0.1.2
     # via markdown-it-py
 moto==5.0.15
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 mypy==1.10.1
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 mypy-boto3-s3==1.35.22
     # via boto3-stubs
 mypy-extensions==1.0.0
@@ -73,13 +73,13 @@ numpy==1.26.4
 packaging==24.0
     # via pytest
 pandas==2.2.2
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 pluggy==1.5.0
     # via pytest
 polars==1.6.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 pyarrow==16.1.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 pyasn1==0.6.0
     # via rsa
 pycparser==2.22
@@ -88,10 +88,10 @@ pygments==2.18.0
     # via rich
 pytest==8.2.1
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   pytest-mock
 pytest-mock==3.14.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 python-dateutil==2.9.0.post0
     # via
     #   botocore
@@ -106,21 +106,21 @@ pyyaml==6.0.1
     #   responses
 requests==2.32.3
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   moto
     #   responses
 responses==0.25.3
     # via moto
 rich==13.7.1
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   rich-click
 rich-click==1.8.2
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 rsa==4.7.2
     # via awscli
 ruff==0.5.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 s3transfer==0.10.1
     # via
     #   awscli
@@ -128,30 +128,32 @@ s3transfer==0.10.1
 six==1.16.0
     # via python-dateutil
 structlog==24.2.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 types-awscrt==0.21.5
     # via botocore-stubs
 types-python-dateutil==2.9.0.20240906
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 types-requests==2.32.0.20240914
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 types-s3transfer==0.10.2
     # via boto3-stubs
 typing-extensions==4.12.0
     # via
+    #   boto3-stubs
     #   mypy
+    #   mypy-boto3-s3
     #   rich-click
 tzdata==2024.1
     # via pandas
 urllib3==2.2.1
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   botocore
     #   requests
     #   responses
     #   types-requests
 us==3.2.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 werkzeug==3.0.4
     # via moto
 xmltodict==0.13.0
