+++ requirements/requirements.txt
@@ -1,9 +1,9 @@
 # This file was autogenerated by uv via the following command:
 #    uv pip compile pyproject.toml -o requirements/requirements.txt
 awscli==1.32.97
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 boto3==1.34.97
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 botocore==1.34.97
     # via
     #   awscli
@@ -15,10 +15,10 @@ charset-normalizer==3.3.2
     # via requests
 click==8.1.7
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   rich-click
 cloudpathlib==0.18.1
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 colorama==0.4.6
     # via awscli
 docutils==0.16
@@ -40,11 +40,11 @@ numpy==1.26.4
     #   pandas
     #   pyarrow
 pandas==2.2.2
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 polars==1.6.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 pyarrow==16.0.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 pyasn1==0.6.0
     # via rsa
 pygments==2.18.0
@@ -58,13 +58,13 @@ pytz==2024.1
 pyyaml==6.0.1
     # via awscli
 requests==2.32.3
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 rich==13.7.1
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   rich-click
 rich-click==1.8.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 rsa==4.7.2
     # via awscli
 s3transfer==0.10.1
@@ -74,15 +74,15 @@ s3transfer==0.10.1
 six==1.16.0
     # via python-dateutil
 structlog==24.1.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
 typing-extensions==4.11.0
     # via rich-click
 tzdata==2024.1
     # via pandas
 urllib3==2.2.1
     # via
-    #   virus-clade-utils (pyproject.toml)
+    #   cladetime (pyproject.toml)
     #   botocore
     #   requests
 us==3.2.0
-    # via virus-clade-utils (pyproject.toml)
+    # via cladetime (pyproject.toml)
