+++ tests/unit/test_get_clade_list.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
-from virus_clade_utils.get_clade_list import main
+from cladetime.get_clade_list import main
 
 
 @pytest.fixture
@@ -33,7 +33,7 @@ def test_clade_list(test_file_path, tmp_path, threshold, weeks, max_clades, expe
     test_genome_metadata = test_file_path / "test_metadata.tsv"
     mock = MagicMock(return_value=test_genome_metadata, name="genome_metadata_download_mock")
 
-    with patch("virus_clade_utils.get_clade_list.download_covid_genome_metadata", mock):
+    with patch("cladetime.get_clade_list.download_covid_genome_metadata", mock):
         actual_list = main("some_bucket", "some_key", tmp_path, threshold, weeks, max_clades)
 
     assert set(expected_list) == set(actual_list)
